Linked up the Document Viewer and the new Worked Examples Viewer
Now every view controller that is unique design wise has a its own view
controller and goes back along the story board to its respective
previous view the user was on, instead of going 'back' to the documents
browser. Currently the passing of data between the views isnt working
correctly this will be fixed later.